Module substrate_test_runtime::substrate_test_pallet::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet. Note that members of this module cannot be used directly and are only provided for documentation purposes.

Functions

bench_call

Legacy call used in transaction pool benchmarks.

call_do_not_propagate

This call is validated as non-propagable ValidTransaction.

call_with_priority

This call is validated as ValidTransaction with given priority.

deposit_log_digest_item

Deposit given digest items into the system storage. They will be included in a header during finalization.

fill_block

Fill the block weight up to the given ratio.

include_data

Implicitly fill a block body with some data.

indexed_call

Create an index for this call.

offchain_index_clear

Remove a key and an associated value from the offchain database.

offchain_index_set

Write a key value pair to the offchain database.

read

Read X times from the state some data.

read_and_panic

Read X times from the state some data and then panic!

storage_change

Put/delete some data from storage. Intended to use as an unsigned extrinsic.