Trait substrate_test_runtime::TestAPI

pub trait TestAPI<Block: BlockT>: Core<Block> {
    fn balance_of(
        &self, 
        at: &BlockId<Block>, 
        id: AccountId
    ) -> Result<u64, Error> { ... }

    fn balance_of_with_context(
        &self, 
        at: &BlockId<Block>, 
        context: ExecutionContext, 
        id: AccountId
    ) -> Result<u64, Error> { ... }

    fn benchmark_add_one(
        &self, 
        at: &BlockId<Block>, 
        val: &u64
    ) -> Result<u64, Error> { ... }

    fn benchmark_add_one_with_context(
        &self, 
        at: &BlockId<Block>, 
        context: ExecutionContext, 
        val: &u64
    ) -> Result<u64, Error> { ... }

    fn benchmark_vector_add_one(
        &self, 
        at: &BlockId<Block>, 
        vec: &Vec<u64>
    ) -> Result<Vec<u64>, Error> { ... }

    fn benchmark_vector_add_one_with_context(
        &self, 
        at: &BlockId<Block>, 
        context: ExecutionContext, 
        vec: &Vec<u64>
    ) -> Result<Vec<u64>, Error> { ... }

    fn fail_convert_parameter(
        &self, 
        at: &BlockId<Block>, 
        param: DecodeFails<Block>
    ) -> Result<(), Error> { ... }

    fn fail_convert_parameter_with_context(
        &self, 
        at: &BlockId<Block>, 
        context: ExecutionContext, 
        param: DecodeFails<Block>
    ) -> Result<(), Error> { ... }

    fn fail_convert_return_value(
        &self, 
        at: &BlockId<Block>
    ) -> Result<DecodeFails<Block>, Error> { ... }

    fn fail_convert_return_value_with_context(
        &self, 
        at: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<DecodeFails<Block>, Error> { ... }

    fn function_signature_changed_before_version_2(
        &self, 
        at: &BlockId<Block>
    ) -> Result<Vec<u64>, Error> { ... }

    fn function_signature_changed_before_version_2_with_context(
        &self, 
        at: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<Vec<u64>, Error> { ... }

    fn function_signature_changed(
        &self, 
        at: &BlockId<Block>
    ) -> Result<u64, Error> { ... }

    fn function_signature_changed_with_context(
        &self, 
        at: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u64, Error> { ... }

    fn fail_on_native(&self, at: &BlockId<Block>) -> Result<u64, Error> { ... }

    fn fail_on_native_with_context(
        &self, 
        at: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u64, Error> { ... }

    fn fail_on_wasm(&self, at: &BlockId<Block>) -> Result<u64, Error> { ... }

    fn fail_on_wasm_with_context(
        &self, 
        at: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u64, Error> { ... }

    fn use_trie(&self, at: &BlockId<Block>) -> Result<u64, Error> { ... }

    fn use_trie_with_context(
        &self, 
        at: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u64, Error> { ... }

    fn benchmark_indirect_call(&self, at: &BlockId<Block>) -> Result<u64, Error> { ... }

    fn benchmark_indirect_call_with_context(
        &self, 
        at: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u64, Error> { ... }

    fn benchmark_direct_call(&self, at: &BlockId<Block>) -> Result<u64, Error> { ... }

    fn benchmark_direct_call_with_context(
        &self, 
        at: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u64, Error> { ... }

    fn returns_mutable_static(&self, at: &BlockId<Block>) -> Result<u64, Error> { ... }

    fn returns_mutable_static_with_context(
        &self, 
        at: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u64, Error> { ... }

    fn allocates_huge_stack_array(
        &self, 
        at: &BlockId<Block>, 
        trap: bool
    ) -> Result<Vec<u8>, Error> { ... }

    fn allocates_huge_stack_array_with_context(
        &self, 
        at: &BlockId<Block>, 
        context: ExecutionContext, 
        trap: bool
    ) -> Result<Vec<u8>, Error> { ... }

    fn vec_with_capacity(
        &self, 
        at: &BlockId<Block>, 
        size: u32
    ) -> Result<Vec<u8>, Error> { ... }

    fn vec_with_capacity_with_context(
        &self, 
        at: &BlockId<Block>, 
        context: ExecutionContext, 
        size: u32
    ) -> Result<Vec<u8>, Error> { ... }

    fn get_block_number(&self, at: &BlockId<Block>) -> Result<u64, Error> { ... }

    fn get_block_number_with_context(
        &self, 
        at: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u64, Error> { ... }

    fn take_block_number(
        &self, 
        at: &BlockId<Block>
    ) -> Result<Option<u64>, Error> { ... }

    fn take_block_number_with_context(
        &self, 
        at: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<Option<u64>, Error> { ... }

    fn without_initialize_block(
        &self, 
        at: &BlockId<Block>
    ) -> Result<bool, Error> { ... }

    fn without_initialize_block_with_context(
        &self, 
        at: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<bool, Error> { ... }

    fn test_ed25519_crypto(
        &self, 
        at: &BlockId<Block>
    ) -> Result<(AppSignature, AppPublic), Error> { ... }

    fn test_ed25519_crypto_with_context(
        &self, 
        at: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<(AppSignature, AppPublic), Error> { ... }

    fn test_sr25519_crypto(
        &self, 
        at: &BlockId<Block>
    ) -> Result<(AppSignature, AppPublic), Error> { ... }

    fn test_sr25519_crypto_with_context(
        &self, 
        at: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<(AppSignature, AppPublic), Error> { ... }

    fn test_storage(&self, at: &BlockId<Block>) -> Result<(), Error> { ... }

    fn test_storage_with_context(
        &self, 
        at: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<(), Error> { ... }
}

Provided methods

fn balance_of(&self, at: &BlockId<Block>, id: AccountId) -> Result<u64, Error>

Return the balance of the given account id.

fn balance_of_with_context(
&self,
at: &BlockId<Block>,
context: ExecutionContext,
id: AccountId
) -> Result<u64, Error>

Return the balance of the given account id.

fn benchmark_add_one(
&self,
at: &BlockId<Block>,
val: &u64
) -> Result<u64, Error>

A benchmark function that adds one to the given value and returns the result.

fn benchmark_add_one_with_context(
&self,
at: &BlockId<Block>,
context: ExecutionContext,
val: &u64
) -> Result<u64, Error>

A benchmark function that adds one to the given value and returns the result.

fn benchmark_vector_add_one(
&self,
at: &BlockId<Block>,
vec: &Vec<u64>
) -> Result<Vec<u64>, Error>

A benchmark function that adds one to each value in the given vector and returns the result.

fn benchmark_vector_add_one_with_context(
&self,
at: &BlockId<Block>,
context: ExecutionContext,
vec: &Vec<u64>
) -> Result<Vec<u64>, Error>

A benchmark function that adds one to each value in the given vector and returns the result.

fn fail_convert_parameter(
&self,
at: &BlockId<Block>,
param: DecodeFails<Block>
) -> Result<(), Error>

A function that always fails to convert a parameter between runtime and node.

fn fail_convert_parameter_with_context(
&self,
at: &BlockId<Block>,
context: ExecutionContext,
param: DecodeFails<Block>
) -> Result<(), Error>

A function that always fails to convert a parameter between runtime and node.

fn fail_convert_return_value(
&self,
at: &BlockId<Block>
) -> Result<DecodeFails<Block>, Error>

A function that always fails to convert its return value between runtime and node.

fn fail_convert_return_value_with_context(
&self,
at: &BlockId<Block>,
context: ExecutionContext
) -> Result<DecodeFails<Block>, Error>

A function that always fails to convert its return value between runtime and node.

fn function_signature_changed_before_version_2(
&self,
at: &BlockId<Block>
) -> Result<Vec<u64>, Error>

Deprecated

A function for that the signature changed in version 2.

fn function_signature_changed_before_version_2_with_context(
&self,
at: &BlockId<Block>,
context: ExecutionContext
) -> Result<Vec<u64>, Error>

Deprecated

A function for that the signature changed in version 2.

fn function_signature_changed(&self, at: &BlockId<Block>) -> Result<u64, Error>

The new signature.

fn function_signature_changed_with_context(
&self,
at: &BlockId<Block>,
context: ExecutionContext
) -> Result<u64, Error>

The new signature.

fn fail_on_native(&self, at: &BlockId<Block>) -> Result<u64, Error>

fn fail_on_native_with_context(
&self,
at: &BlockId<Block>,
context: ExecutionContext
) -> Result<u64, Error>

fn fail_on_wasm(&self, at: &BlockId<Block>) -> Result<u64, Error>

fn fail_on_wasm_with_context(
&self,
at: &BlockId<Block>,
context: ExecutionContext
) -> Result<u64, Error>

fn use_trie(&self, at: &BlockId<Block>) -> Result<u64, Error>

trie no_std testing

fn use_trie_with_context(
&self,
at: &BlockId<Block>,
context: ExecutionContext
) -> Result<u64, Error>

trie no_std testing

fn benchmark_indirect_call(&self, at: &BlockId<Block>) -> Result<u64, Error>

fn benchmark_indirect_call_with_context(
&self,
at: &BlockId<Block>,
context: ExecutionContext
) -> Result<u64, Error>

fn benchmark_direct_call(&self, at: &BlockId<Block>) -> Result<u64, Error>

fn benchmark_direct_call_with_context(
&self,
at: &BlockId<Block>,
context: ExecutionContext
) -> Result<u64, Error>

fn returns_mutable_static(&self, at: &BlockId<Block>) -> Result<u64, Error>

fn returns_mutable_static_with_context(
&self,
at: &BlockId<Block>,
context: ExecutionContext
) -> Result<u64, Error>

fn allocates_huge_stack_array(
&self,
at: &BlockId<Block>,
trap: bool
) -> Result<Vec<u8>, Error>

fn allocates_huge_stack_array_with_context(
&self,
at: &BlockId<Block>,
context: ExecutionContext,
trap: bool
) -> Result<Vec<u8>, Error>

fn vec_with_capacity(
&self,
at: &BlockId<Block>,
size: u32
) -> Result<Vec<u8>, Error>

fn vec_with_capacity_with_context(
&self,
at: &BlockId<Block>,
context: ExecutionContext,
size: u32
) -> Result<Vec<u8>, Error>

fn get_block_number(&self, at: &BlockId<Block>) -> Result<u64, Error>

Returns the initialized block number.

fn get_block_number_with_context(
&self,
at: &BlockId<Block>,
context: ExecutionContext
) -> Result<u64, Error>

Returns the initialized block number.

fn take_block_number(&self, at: &BlockId<Block>) -> Result<Option<u64>, Error>

Takes and returns the initialized block number.

fn take_block_number_with_context(
&self,
at: &BlockId<Block>,
context: ExecutionContext
) -> Result<Option<u64>, Error>

Takes and returns the initialized block number.

fn without_initialize_block(&self, at: &BlockId<Block>) -> Result<bool, Error>

Returns if no block was initialized.

fn without_initialize_block_with_context(
&self,
at: &BlockId<Block>,
context: ExecutionContext
) -> Result<bool, Error>

Returns if no block was initialized.

fn test_ed25519_crypto(
&self,
at: &BlockId<Block>
) -> Result<(AppSignature, AppPublic), Error>

Test that ed25519 crypto works in the runtime.

Returns the signature generated for the message ed25519 and the public key.

fn test_ed25519_crypto_with_context(
&self,
at: &BlockId<Block>,
context: ExecutionContext
) -> Result<(AppSignature, AppPublic), Error>

Test that ed25519 crypto works in the runtime.

Returns the signature generated for the message ed25519 and the public key.

fn test_sr25519_crypto(
&self,
at: &BlockId<Block>
) -> Result<(AppSignature, AppPublic), Error>

Test that sr25519 crypto works in the runtime.

Returns the signature generated for the message sr25519.

fn test_sr25519_crypto_with_context(
&self,
at: &BlockId<Block>,
context: ExecutionContext
) -> Result<(AppSignature, AppPublic), Error>

Test that sr25519 crypto works in the runtime.

Returns the signature generated for the message sr25519.

fn test_storage(&self, at: &BlockId<Block>) -> Result<(), Error>

Run various tests against storage.

fn test_storage_with_context(
&self,
at: &BlockId<Block>,
context: ExecutionContext
) -> Result<(), Error>

Run various tests against storage.

Loading content...