Crate substrate_test_runtime

The Substrate runtime. This can be compiled with #[no_std], ready for Wasm.

Re-exports

pub use extrinsic::ExtrinsicBuilder;

pub use extrinsic::Transfer;

Modules

api

currency

extrinsic

Provides utils for building the Extrinsic instances used with substrate-test-runtime.

genesismap

Tool for creating the genesis block.

storage_key_generator

Some tests require the hashed keys of the storage. As the values of hashed keys are not trivial to guess, this small module provides the values of the keys, and the code which is required to generate the keys.

substrate_test_pallet

substrate-test pallet

wasm_binary_logging_disabled

Structs

BabeEpochConfiguration

Configuration data used by the BABE consensus engine that may change with epochs.

BlockHashCount

CheckSubstrateCall

EpochDuration

ExistentialDeposit

H256

Fixed-size uninterpreted hash type with 32 bytes (256 bits) size.

MaxLocks

MaxReserves

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SessionKeys

Slot

Unit type wrapper that represents a slot.

TransferData

Transfer data extracted from Extrinsic containing Balances::transfer_allow_death.

Version

Enums

AllowedSlots

Types of allowed slots.

BalancesCall

Contains a variant per dispatchable extrinsic that this pallet has.

OriginCaller

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

RuntimeViewFunction

Runtime query type.

UtilityCall

Contains a variant per dispatchable extrinsic that this pallet has.

Constants

RUNTIME_API_VERSIONS

TEST_OWNER

The test owner to test proof of possession generation and verification for the session keys

TEST_RUNTIME_BABE_EPOCH_CONFIGURATION

VERSION

Test runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

WASM_BINARY_PATH

Traits

TestAPI

Functions

native_version

Native version.

wasm_binary_logging_disabled_unwrap

Wasm binary unwrapped. If built with SKIP_WASM_BUILD, the function panics.

wasm_binary_unwrap

Wasm binary unwrapped. If built with SKIP_WASM_BUILD, the function panics.

Type Aliases

AccountId

An identifier for an account on this system.

Address

The address format for describing accounts.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AuraId

Babe

BabeConfig

Balance

Balance of an account.

Balances

BalancesConfig

Block

A test block.

BlockNumber

The block number type used in this runtime.

Bls381Pop

Bls381Public

Digest

The digest of a block.

DigestItem

The item of a block digest.

EcdsaBls381Pop

EcdsaBls381Public

Executive

Extrinsic

Unchecked extrinsic type as expected by this runtime.

Hash

A simple hash type for all our hashing.

Hashing

The hashing algorithm used.

Header

A test block’s header.

Nonce

Index of a transaction.

Pair

Signature

SignedPayload

The payload being signed in transactions.

SubstrateTest

SubstrateTestConfig

System

SystemConfig

TxExtension

The extension to the basic transaction logic.

Utility