Module substrate_test_runtime::system

System manager: Handles all of the top-level stuff; executing block/transaction, setting code and depositing logs.

Structs

BlockExecutor

The block executor.

GenesisConfig

Can be used to configure the genesis state of this pallet.

Pallet

The pallet implementing the on-chain logic.

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Traits

Config

Configuration trait of this pallet.

Functions

authorities

balance_of

balance_of_key

execute_block

execute_transaction

Execute a transaction outside of the block execution function. This doesn’t attempt to validate anything regarding the block.

finalize_block

Finalize the block.

get_block_number

initialize_block

nonce_of

polish_block

Actually execute all transitioning for block.

take_block_number

validate_transaction

Execute a transaction outside of the block execution function. This doesn’t attempt to validate anything regarding the block.

Type Definitions

Authorities

ExtrinsicData

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

NewAuthorities

Number

ParentHash

StorageDigest