Module substrate_test_runtime::system

System manager: Handles all of the top-level stuff; executing block/transaction, setting code and depositing logs.

Re-exports

pub use tt_error_token_0 as tt_error_token;

Structs

BlockExecutor

The block executor.

GenesisConfig

Genesis config for the module, allow to build genesis storage.

Module

Enums

Call

Dispatchable calls.

Functions

authorities

balance_of

balance_of_key

execute_block

execute_transaction

Execute a transaction outside of the block execution function. This doesn’t attempt to validate anything regarding the block.

finalize_block

Finalize the block.

get_block_number

initialize_block

nonce_of

polish_block

Actually execute all transitioning for block.

take_block_number

validate_transaction

Execute a transaction outside of the block execution function. This doesn’t attempt to validate anything regarding the block.

Type Definitions

Pallet

Type alias to Module, to be used by construct_runtime.