Trait substrate_test_runtime::TestAPI

pub trait TestAPI<Block: BlockT>: Core<Block> {

    fn balance_of(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        id: AccountId
    ) -> Result<u64, Self::Error> { ... }

    fn balance_of_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        id: AccountId
    ) -> Result<u64, Self::Error> { ... }

    fn benchmark_add_one(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        val: &u64
    ) -> Result<u64, Self::Error> { ... }

    fn benchmark_add_one_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        val: &u64
    ) -> Result<u64, Self::Error> { ... }

    fn benchmark_vector_add_one(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        vec: &Vec<u64>
    ) -> Result<Vec<u64>, Self::Error> { ... }

    fn benchmark_vector_add_one_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        vec: &Vec<u64>
    ) -> Result<Vec<u64>, Self::Error> { ... }

    fn fail_convert_parameter(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        param: DecodeFails<Block>
    ) -> Result<(), Self::Error> { ... }

    fn fail_convert_parameter_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        param: DecodeFails<Block>
    ) -> Result<(), Self::Error> { ... }

    fn fail_convert_return_value(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<DecodeFails<Block>, Self::Error> { ... }

    fn fail_convert_return_value_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<DecodeFails<Block>, Self::Error> { ... }

    fn function_signature_changed_before_version_2(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<Vec<u64>, Self::Error> { ... }

    fn function_signature_changed_before_version_2_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<Vec<u64>, Self::Error> { ... }

    fn function_signature_changed(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<u64, Self::Error> { ... }

    fn function_signature_changed_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u64, Self::Error> { ... }

    fn fail_on_native(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<u64, Self::Error> { ... }

    fn fail_on_native_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u64, Self::Error> { ... }

    fn fail_on_wasm(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<u64, Self::Error> { ... }

    fn fail_on_wasm_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u64, Self::Error> { ... }

    fn use_trie(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<u64, Self::Error> { ... }

    fn use_trie_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u64, Self::Error> { ... }

    fn benchmark_indirect_call(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<u64, Self::Error> { ... }

    fn benchmark_indirect_call_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u64, Self::Error> { ... }

    fn benchmark_direct_call(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<u64, Self::Error> { ... }

    fn benchmark_direct_call_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u64, Self::Error> { ... }

    fn vec_with_capacity(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        size: u32
    ) -> Result<Vec<u8>, Self::Error> { ... }

    fn vec_with_capacity_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        size: u32
    ) -> Result<Vec<u8>, Self::Error> { ... }

    fn get_block_number(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<u64, Self::Error> { ... }

    fn get_block_number_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u64, Self::Error> { ... }

    fn take_block_number(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<Option<u64>, Self::Error> { ... }

    fn take_block_number_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<Option<u64>, Self::Error> { ... }

    fn without_initialize_block(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<bool, Self::Error> { ... }

    fn without_initialize_block_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<bool, Self::Error> { ... }

    fn test_ed25519_crypto(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<(AppSignature, AppPublic), Self::Error> { ... }

    fn test_ed25519_crypto_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<(AppSignature, AppPublic), Self::Error> { ... }

    fn test_sr25519_crypto(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<(AppSignature, AppPublic), Self::Error> { ... }

    fn test_sr25519_crypto_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<(AppSignature, AppPublic), Self::Error> { ... }

    fn test_ecdsa_crypto(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<(AppSignature, AppPublic), Self::Error> { ... }

    fn test_ecdsa_crypto_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<(AppSignature, AppPublic), Self::Error> { ... }

    fn test_storage(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<(), Self::Error> { ... }

    fn test_storage_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<(), Self::Error> { ... }

    fn test_witness(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        proof: StorageProof, 
        root: Hash
    ) -> Result<(), Self::Error> { ... }

    fn test_witness_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        proof: StorageProof, 
        root: Hash
    ) -> Result<(), Self::Error> { ... }

    fn test_multiple_arguments(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        data: Vec<u8>, 
        other: Vec<u8>, 
        num: u32
    ) -> Result<(), Self::Error> { ... }

    fn test_multiple_arguments_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        data: Vec<u8>, 
        other: Vec<u8>, 
        num: u32
    ) -> Result<(), Self::Error> { ... }

    fn do_trace_log(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<(), Self::Error> { ... }

    fn do_trace_log_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<(), Self::Error> { ... }

}

Provided methods

fn balance_of(
&self,
__runtime_api_at_param__: &BlockId<Block>,
id: AccountId
) -> Result<u64, Self::Error>

Return the balance of the given account id.

fn balance_of_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
id: AccountId
) -> Result<u64, Self::Error>

Return the balance of the given account id.

fn benchmark_add_one(
&self,
__runtime_api_at_param__: &BlockId<Block>,
val: &u64
) -> Result<u64, Self::Error>

A benchmark function that adds one to the given value and returns the result.

fn benchmark_add_one_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
val: &u64
) -> Result<u64, Self::Error>

A benchmark function that adds one to the given value and returns the result.

fn benchmark_vector_add_one(
&self,
__runtime_api_at_param__: &BlockId<Block>,
vec: &Vec<u64>
) -> Result<Vec<u64>, Self::Error>

A benchmark function that adds one to each value in the given vector and returns the result.

fn benchmark_vector_add_one_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
vec: &Vec<u64>
) -> Result<Vec<u64>, Self::Error>

A benchmark function that adds one to each value in the given vector and returns the result.

fn fail_convert_parameter(
&self,
__runtime_api_at_param__: &BlockId<Block>,
param: DecodeFails<Block>
) -> Result<(), Self::Error>

A function that always fails to convert a parameter between runtime and node.

fn fail_convert_parameter_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
param: DecodeFails<Block>
) -> Result<(), Self::Error>

A function that always fails to convert a parameter between runtime and node.

fn fail_convert_return_value(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<DecodeFails<Block>, Self::Error>

A function that always fails to convert its return value between runtime and node.

fn fail_convert_return_value_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<DecodeFails<Block>, Self::Error>

A function that always fails to convert its return value between runtime and node.

fn function_signature_changed_before_version_2(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<Vec<u64>, Self::Error>

👎 Deprecated

A function for that the signature changed in version 2.

fn function_signature_changed_before_version_2_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<Vec<u64>, Self::Error>

👎 Deprecated

A function for that the signature changed in version 2.

fn function_signature_changed(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<u64, Self::Error>

The new signature.

fn function_signature_changed_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<u64, Self::Error>

The new signature.

fn fail_on_native(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<u64, Self::Error>

fn fail_on_native_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<u64, Self::Error>

fn fail_on_wasm(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<u64, Self::Error>

fn fail_on_wasm_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<u64, Self::Error>

fn use_trie(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<u64, Self::Error>

trie no_std testing

fn use_trie_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<u64, Self::Error>

trie no_std testing

fn benchmark_indirect_call(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<u64, Self::Error>

fn benchmark_indirect_call_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<u64, Self::Error>

fn benchmark_direct_call(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<u64, Self::Error>

fn benchmark_direct_call_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<u64, Self::Error>

fn vec_with_capacity(
&self,
__runtime_api_at_param__: &BlockId<Block>,
size: u32
) -> Result<Vec<u8>, Self::Error>

fn vec_with_capacity_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
size: u32
) -> Result<Vec<u8>, Self::Error>

fn get_block_number(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<u64, Self::Error>

Returns the initialized block number.

fn get_block_number_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<u64, Self::Error>

Returns the initialized block number.

fn take_block_number(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<Option<u64>, Self::Error>

Takes and returns the initialized block number.

fn take_block_number_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<Option<u64>, Self::Error>

Takes and returns the initialized block number.

fn without_initialize_block(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<bool, Self::Error>

Returns if no block was initialized.

fn without_initialize_block_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<bool, Self::Error>

Returns if no block was initialized.

fn test_ed25519_crypto(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<(AppSignature, AppPublic), Self::Error>

Test that ed25519 crypto works in the runtime.

Returns the signature generated for the message ed25519 and the public key.

fn test_ed25519_crypto_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<(AppSignature, AppPublic), Self::Error>

Test that ed25519 crypto works in the runtime.

Returns the signature generated for the message ed25519 and the public key.

fn test_sr25519_crypto(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<(AppSignature, AppPublic), Self::Error>

Test that sr25519 crypto works in the runtime.

Returns the signature generated for the message sr25519.

fn test_sr25519_crypto_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<(AppSignature, AppPublic), Self::Error>

Test that sr25519 crypto works in the runtime.

Returns the signature generated for the message sr25519.

fn test_ecdsa_crypto(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<(AppSignature, AppPublic), Self::Error>

Test that ecdsa crypto works in the runtime.

Returns the signature generated for the message ecdsa.

fn test_ecdsa_crypto_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<(AppSignature, AppPublic), Self::Error>

Test that ecdsa crypto works in the runtime.

Returns the signature generated for the message ecdsa.

fn test_storage(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<(), Self::Error>

Run various tests against storage.

fn test_storage_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<(), Self::Error>

Run various tests against storage.

fn test_witness(
&self,
__runtime_api_at_param__: &BlockId<Block>,
proof: StorageProof,
root: Hash
) -> Result<(), Self::Error>

Check a witness.

fn test_witness_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
proof: StorageProof,
root: Hash
) -> Result<(), Self::Error>

Check a witness.

fn test_multiple_arguments(
&self,
__runtime_api_at_param__: &BlockId<Block>,
data: Vec<u8>,
other: Vec<u8>,
num: u32
) -> Result<(), Self::Error>

Test that ensures that we can call a function that takes multiple arguments.

fn test_multiple_arguments_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
data: Vec<u8>,
other: Vec<u8>,
num: u32
) -> Result<(), Self::Error>

Test that ensures that we can call a function that takes multiple arguments.

fn do_trace_log(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<(), Self::Error>

Traces log “Hey I’m runtime.”

fn do_trace_log_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<(), Self::Error>

Traces log “Hey I’m runtime.”